Fix 'compute' reduction with primitive Int63
Using compute on a term like (1+1)%int63 was generating an anomaly
because the cbv_value's INT i and VAR (0, mkInt i) were both representing
a primitive integer but in incompatible ways.

My fix uses VAR (0, mkInt i) everywhere to remove this difference thus
solving the problem.

Note that I added the subdir arithmetic in the test-suite, and it now
passes the tests.