Patch CBV reduction of primitive CONSTR results
The CBV reduction expects all arguments of a CONSTR on the stack to
continue its reduction (e.g. choose the branche of a CASE), but the
primitive operations were encoding the arguments inside the term.

This prevented correct reduction of many lambda terms using 'let' or
'match' when using 'cbv' or 'compute' tactics. Terms were either half reduced
or generating an anomaly because of failed reduction.

This commit patch this bug by always extracting all the arguments of CONSTR
terms after primitive computations and place them on the top of the
stack.

It also adds a test case generating the anomaly on older versions.