Add Floats to standard library
All supported floating point operations are defined on emulated floats.
Then we register the primitive type and functions, and add conversion
functions to and from the emulated type.
Finally we put axioms to state that primitive operations behave exactly
the same as emulated operations.

CREDITS: Most of the code inside EmulatedFloat is adapted from the Flocq
library.

NOTE: For the moment this code will not compile if native compilation is
enabled in the configuration phase. This will be resolved later when
native_compute will be supported by primitive floats.