Add primitive floats to 'vm_compute'
This commit add float instructions to the VM, their encoding in bytecode
and the interpretation of primitive float values after the reduction.

Tests were updated to also include VM computations and check for double
rounding.

The flag '-std=c99' is added to the C compiler flags to ensure that float
computation strictly follows the norm (ie. i387 80-bits format is not
used as an optimization).